requirements.txt
CherryPy >= 17.3.0
wget >=3.2
pyserial == 3.4